refactor(*) adjusting encoding's package directory
