+++ README_CN.md
@@ -335,7 +335,7 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/yomorun/y3-codec-golang/pkg/spec/encoding"
+	"github.com/yomorun/y3-codec-golang/pkg/encoding"
 )
 
 func main() {
