+++ pkg/encoding/nvarint.go
