+++ pkg/encoding/varfloat.go
