+++ pkg/encoding/varfloat_test.go
